Reorganize pt 1 - Move cli script definitions into index
