+++ packages/iobio-react-components/src/utils/statisticsClient/statisticsTools.mts
@@ -21,7 +21,6 @@
 
 import { DataBroker } from 'iobio-charts/data_broker.js';
 import fs from 'node:fs';
-import readline from 'node:readline/promises';
 import { calculateMeanCoverage, getBamStatistics } from '../iobioTools.mts';
 import { type StatsOutput } from '../iobioTypes.ts';
 
@@ -38,17 +37,18 @@ export const generateIobioStats = async ({
 	fileUrl,
 	fileName,
 	indexFileUrl = '',
+	serverUrl = '',
 	enableFileOutput = false,
 	onComplete,
 }: {
 	fileUrl: string;
 	fileName?: string;
 	indexFileUrl?: string | null;
+	serverUrl?: string;
 	enableFileOutput?: boolean;
 	onComplete?: CompleteCallback | null;
 }): Promise<void> => {
 	// Generate Statistics
-	const serverUrl = process.env.IOBIO_SERVER_URL;
 	const db = new DataBroker(fileUrl, { server: serverUrl });
 	if (indexFileUrl) db.indexUrl = indexFileUrl;
 
@@ -102,23 +102,4 @@ export const outputFile = (fileName: string, fileData: StatsOutput): void => {
 			console.log(`\nFile output to ${fileName}\n`);
 		}
 	});
-};
-
-/** Launch command line prompt for user input, then generate stats */
-export const statisticsCLI = async (): Promise<void> => {
-	const readlineInterface = readline.createInterface({
-		input: process.stdin,
-		output: process.stdout,
-	});
-	console.log('***** Overture Components: Iobio Metadata Generator *****');
-
-	const fileUrl = await readlineInterface.question('\nBam File URL: ');
-	if (!fileUrl) throw new Error('Alignment URL is required to generate statistics \nusage: pnpm run stats ${url}');
-
-	const indexFileUrl = await readlineInterface.question('\nIndex File URL (optional): ');
-	const outputOption = await readlineInterface.question('\nOutput as JSON? (Y/N): ');
-	const enableFileOutput = outputOption.toLowerCase() === 'y';
-	readlineInterface.close();
-
-	generateIobioStats({ fileUrl, indexFileUrl, enableFileOutput });
 };
